ExtraRibs property is list of floats from 0.0 to 1.0. For each float in the list an additional rib is added. So, if you need an extra rib in one place, but don't want to have to add a bunch of extra ribs to get one in that spot you can do it with this. For example, if you have Items = 4, you get ribs at 0.0, 0.3333, 0.6667, 1.0. Now, let's say you want ribs at 0.2 and 0.5 in addition to those 4. To get that you add in the ExtraRibs float list the values 0.2 and 0.5.
At first I was going to make this take the place of the evenly spaced ribs using Items, but then decided to make it extra in addition to those. The user can set Items to 0 if he only wants these.

Feel free to modify this code in any way you see proper or to not use it all if you don't think is right for your workbench.